requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
Django==5.2.4
django-ninja==1.4.3
freezegun==1.5.3
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
ruff==0.12.4
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
